refactor: changed from chrome-services to browser-services
