+++ entrypoints/popup/main/popup_body.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
-import { exchangeRefreshForAccessToken, isAccessTokenValid } from '../../../utils/chrome-services';
-import { log } from '../../../utils/logger';
-import { useActions, useLoaderStore, useUser } from '../../../utils/zustand';
+import { exchangeRefreshForAccessToken, isAccessTokenValid } from '@/utils/browser-services';
+import { log } from '@/utils';
+import { useActions, useLoaderStore, useUser } from '@/utils/zustand';
 import { GoToSheet, Login } from './';
