+++ entrypoints/popup/options/Logout.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import { revokeTokens } from "../../../utils/chrome-services";
-import { log } from '../../../utils/logger';
-import { useActions, useUser } from '../../../utils/zustand';
+import { revokeTokens } from "@/utils/browser-services";
+import { log } from '@/utils';
+import { useActions, useUser } from '@/utils/zustand';
 
 export const Logout = () => {
     const [style, setStyle] = useState("");
